refactor: 테스트코드 Owner 등 임시 임포트 변경 및 삭제
